chore: code-splitting in the FirmwareModal, separation of steps to isolated files
